move querying helper scopes into metabase-query lib under sdk
