+++ frontend/src/metabase/embedding-sdk/metabase-query/metabase-lib-query-builder.ts
@@ -6,7 +6,7 @@ import {
   getMetricSourceTableId,
   getTableDatabaseId,
   getTableId,
-} from "./guards";
+} from "./accessors";
 import {
   applyAggregations,
   applyBreakouts,
