+++ frontend/src/metabase/embedding-sdk/metabase-query/metric-query-builder.ts
@@ -11,6 +11,8 @@ import {
   getMetricId,
   getMetricSourceCardId,
   getMetricSourceTableId,
+} from "./accessors";
+import {
   isFieldSchema,
   isMeasureSchema,
   isSegmentSchema,
