+++ frontend/src/metabase/embedding-sdk/metabase-query/runtime-types.ts
@@ -1,24 +1,10 @@
+import type { FilterOperator as LibFilterOperator } from "metabase-lib/common";
+
 import type { FieldSchema } from "./schema";
 
 export type ID = string | number;
 
-export type FilterOperator =
-  | "="
-  | "!="
-  | ">"
-  | ">="
-  | "<"
-  | "<="
-  | "between"
-  | "contains"
-  | "does-not-contain"
-  | "starts-with"
-  | "ends-with"
-  | "is-empty"
-  | "not-empty"
-  | "is-null"
-  | "not-null"
-  | "time-interval";
+export type FilterOperator = LibFilterOperator | "time-interval";
 
 export type QuestionQueryRuntime = {
   questionId: ID;
