+++ frontend/src/metabase/embedding-sdk/metabase-query/validation.ts
@@ -1,5 +1,5 @@
+import { getMetricMappedTableIds } from "./accessors";
 import {
-  getMetricMappedTableIds,
   isCountAggregation,
   isFieldAggregation,
   isMeasureSchema,
